Release accessing read function of observable
Tracked down a memory leak in my application that appears to originate
from the line. I was not able to reproduce this with core knockback and
it's dependencies. I was also not able to run the test suite on windows.
This is also my first commit to github. Given all that I won't be
offended by a rejection ;)

Thinking through it, there doesn't appear to be any reason why the
release function shouldn't be using peek here, but I can't prove that it
causes a problem outside of my application.

If you'd like more information about the approach we took with the app
I'd be happy to explain and share some code samples.

Also just wanted to say thanks, using knockback has helped greatly
reduce the amount of custom javascript code that I've seen backbone only
projects produce.